Move the Repository linter to use Gateway
Still checks ROM::Repository for instantiation, and still references as
Repository in linter mixins
