+++ lib/rom/lint/spec.rb
@@ -1,8 +1,8 @@
-require 'rom/lint/repository'
+require 'rom/lint/gateway'
 require 'rom/lint/enumerable_dataset'
 
 RSpec.shared_examples "a rom repository" do
-  ROM::Lint::Repository.each_lint do |name, linter|
+  ROM::Lint::Gateway.each_lint do |name, linter|
     it name do
       result = linter.new(identifier, repository, uri).lint(name)
       expect(result).to be_truthy
